refactor: extract GridControls styles to CSS modules and component
- Extracted inline styles from `GridControls.tsx` to `GridControls.module.css`
- Moved `TableSelector` component to its own file `TableSelector.tsx`
- Updated `GridControls.tsx` to use the new component and CSS module
- Verified functionality with unit tests and manual frontend verification script
