+++ packages/osdlabel/src/components/GridControls.module.css
@@ -1,0 +1,61 @@
+.container {
+  position: relative;
+}
+
+.triggerButton {
+  padding: 4px 8px;
+  border: 1px solid #444;
+  border-radius: 4px;
+  background: #333;
+  color: #fff;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  gap: 6px;
+  font-size: 12px;
+}
+
+.popover {
+  position: absolute;
+  top: 100%;
+  left: 0;
+  padding-top: 4px;
+  pointer-events: auto;
+  min-width: 100px;
+  z-index: 1000;
+}
+
+.popoverContent {
+  background: #2a2a2a;
+  border: 1px solid #444;
+  border-radius: 4px;
+  padding: 8px;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.3);
+}
+
+.dimensionsLabel {
+  margin-bottom: 6px;
+  font-size: 11px;
+  color: #aaa;
+  text-align: center;
+}
+
+.gridContainer {
+  display: grid;
+  gap: 2px;
+}
+
+.gridCell {
+  width: 18px;
+  height: 18px;
+  border: 1px solid #555;
+  cursor: pointer;
+  border-radius: 2px;
+  transition: background 0.1s ease;
+}
+
+.controlsContainer {
+  display: flex;
+  align-items: center;
+  font-family: system-ui, sans-serif;
+}
